requirements.txt
# Python dependencies for the Wing Commander Saga to Godot Converter

# Core dependencies
pydantic>=2.11.7
PyYAML>=6.0.2
networkx>=3.5
watchdog>=6.0.0
langgraph==0.6.6

# Development dependencies
pytest==8.4.1
pytest-cov>=6.0.0
pytest-mock>=3.14.1
black>=25.1.0
flake8>=7.3.0
mypy>=1.17.1
isort>=6.0.1
